// src/module/common/momo/momo-interface.ts
export interface CreateMomoData {
  partnerCode: string;
  accessKey: string;
  secretKey: string;
  requestId: string;
  orderInfo: string;
  orderId: string;
  redirectUrl: string;
  ipnUrl: string;
  amount: string;
  requestType: string;
  extraData: string;
  signature: string;
  lang: string;
}